requirements.txt
asgiref==3.8.1
coverage==7.5.4
Django==4.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
typing_extensions==4.12.2
